Feature: TestPG 연동 로직 작성 완료 및 테스트 코드 작성 및 테스트 완료
minor_fix:
- linting 수정
- 헥사고날아키텍쳐에 맞게 파일 위치 변경
- TestPg 응답으로 오는 일부 필드에 대한 자료형 변경

+++ modules/external/pg-client/src/main/kotlin/im/bigs/pg/external/pg/dto/TestPGResponse.kt
@@ -1,11 +1,9 @@
 package im.bigs.pg.external.pg.dto
 
-import java.math.BigDecimal
-
 data class TestPGResponse(
     val approvalCode: String,
     val approvedAt: String,
     val maskedCardLast4: String,
-    val amount: BigDecimal,
+    val amount: Int,
     val status: String
 )
